+++ tests/unittest/_torch/auto_deploy/unit/singlegpu/custom_ops/test_mamba_rms_norm.py
@@ -1,8 +1,10 @@
 import pytest
 import torch
 
-import tensorrt_llm._torch.auto_deploy.custom_ops  # noqa: F401
-from tensorrt_llm._torch.auto_deploy.models.patches.nemotron_h import _rms_norm_ref
+from tensorrt_llm._torch.auto_deploy.custom_ops.rms_norm import (
+    gated_rms_norm_ref,
+    triton_rmsnorm_gated,
+)
 
 
 @pytest.mark.skipif(
@@ -19,12 +21,12 @@ def test_custom_op_matches_ref(B, T, H, group, use_gate, dtype):
     z = torch.randn_like(x) if use_gate else None
     w = torch.ones(H, dtype=dtype, device=device)
 
-    y_ref = _rms_norm_ref(
+    y_ref = gated_rms_norm_ref(
         x, w, bias=None, z=z, eps=1e-5, group_size=group, norm_before_gate=False, upcast=True
     )
 
     # Custom op (currently returns fp32). Cast it back to x.dtype for apples-to-apples with ref.
-    y_op_fp32 = torch.ops.auto_deploy.triton_rmsnorm_gated(x, w, z, 1e-5, group, False)
+    y_op_fp32 = triton_rmsnorm_gated(x, w, z, 1e-5, group, False)
     y_op = y_op_fp32.to(x.dtype)
 
     assert y_ref.dtype == x.dtype and y_op.dtype == x.dtype
